backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
chromadb==0.6.3
google-generativeai==0.8.4
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.10.4
apscheduler==3.10.4
lxml==5.3.0
